socket working

+++ app/src/main/java/com/example/chaya/bontact/Data/InnerConversationProvider.java
@@ -58,8 +58,8 @@ public class InnerConversationProvider extends ContentProvider {
 
     @Override
     public int delete(Uri uri, String selection, String[] selectionArgs) {
-        int result= dbBontact.delete(Contract.InnerConversation.TABLE_NAME, selection,selectionArgs);
-        return  result;
+        long result= dbBontact.delete(Contract.InnerConversation.TABLE_NAME, selection,selectionArgs);
+        return  (int) result;
     }
 
     @Override
